+++ src/fitnessTracker.ts
@@ -1,22 +1,9 @@
 import { NativeModules } from 'react-native';
 import DeviceInfo from 'react-native-device-info';
 
+import { IStepTrackerData, IStepTrackerStatus, IWeekDailySteps } from './types';
+
 const { RNFitnessTracker } = NativeModules;
-
-export interface IStepTrackerStatus {
-  authorized: boolean;
-  shouldOpenAppSettings: boolean;
-  trackingNotSupported?: boolean;
-}
-
-export interface IWeekDailySteps {
-  [key: string]: number;
-}
-
-export interface IStepTrackerData {
-  stepsToday: number;
-  stepsThisWeek: IWeekDailySteps;
-}
 
 const iosAuthorizationStatusCheck = (status: string): IStepTrackerStatus => {
   if (status === 'authorized') {
@@ -32,7 +19,7 @@ const iosAuthorizationStatusCheck = (status: string): IStepTrackerStatus => {
  * `iOS only!` returns if step tracking is supported on device
  * @return {Promise<boolean>}
  */
-export const isStepTrackingSupported = (): Promise<boolean> =>
+const isStepTrackingSupported = (): Promise<boolean> =>
   new Promise(resolve => {
     RNFitnessTracker.isStepTrackingSupported((available: number) => {
       resolve(available ? true : false);
@@ -43,7 +30,7 @@ export const isStepTrackingSupported = (): Promise<boolean> =>
  * Returns if step tracking is authorized and available on `Android`
  * @return {Promise<IStepTrackerStatus>}
  */
-export const isStepTrackingAvailableAndroid = (): Promise<IStepTrackerStatus> =>
+const isStepTrackingAvailableAndroid = (): Promise<IStepTrackerStatus> =>
   new Promise(resolve => {
     RNFitnessTracker.authorize((authorized: boolean) => {
       resolve({ authorized, shouldOpenAppSettings: false });
@@ -54,7 +41,7 @@ export const isStepTrackingAvailableAndroid = (): Promise<IStepTrackerStatus> =>
  * Returns if step tracking is authorized and available on `iOS`
  * @return {Promise<IStepTrackerStatus>}
  */
-export const isStepTrackingAvailableIOS = (): Promise<IStepTrackerStatus> =>
+const isStepTrackingAvailableIOS = (): Promise<IStepTrackerStatus> =>
   new Promise(resolve => {
     RNFitnessTracker.isAuthorizedToUseCoreMotion((status: string) => {
       resolve(iosAuthorizationStatusCheck(status));
@@ -65,7 +52,7 @@ export const isStepTrackingAvailableIOS = (): Promise<IStepTrackerStatus> =>
  * Returns if step tracking is authorized and available on both platforms
  * @return {Promise<IStepTrackerStatus>}
  */
-export const isStepTrackingAvailable = (): Promise<IStepTrackerStatus> =>
+const isStepTrackingAvailable = (): Promise<IStepTrackerStatus> =>
   new Promise(resolve => {
     if (global.isIOS) {
       RNFitnessTracker.isAuthorizedToUseCoreMotion((status: string) => {
@@ -83,7 +70,7 @@ export const isStepTrackingAvailable = (): Promise<IStepTrackerStatus> =>
  * not supported iOS devices also return `trackingNotSupported: true` param inside the status object
  * @return {Promise<IStepTrackerStatus>}
  */
-export const setupStepTracking = (): Promise<IStepTrackerStatus> =>
+const setupStepTracking = (): Promise<IStepTrackerStatus> =>
   new Promise(resolve => {
     RNFitnessTracker.authorize((authorized: boolean) => {
       if (!global.isIOS) {
@@ -108,7 +95,7 @@ export const setupStepTracking = (): Promise<IStepTrackerStatus> =>
  * Returns number of steps today
  * @return {Promise<number>}
  */
-export const getStepsToday = (): Promise<number> =>
+const getStepsToday = (): Promise<number> =>
   new Promise(resolve => {
     RNFitnessTracker.getStepsToday((steps: number) => {
       resolve(steps);
@@ -119,7 +106,7 @@ export const getStepsToday = (): Promise<number> =>
  * Returns number of steps this week
  * @return {Promise<Number>}
  */
-export const getStepsThisWeek = (): Promise<number> =>
+const getStepsThisWeek = (): Promise<number> =>
   new Promise(resolve => {
     RNFitnessTracker.getWeekData((steps: number) => {
       resolve(steps);
@@ -130,7 +117,7 @@ export const getStepsThisWeek = (): Promise<number> =>
  * Returns weekly steps object
  * @return {Promise<IWeekDailySteps>}
  */
-export const getWeeklySteps = (): Promise<IWeekDailySteps> =>
+const getWeeklySteps = (): Promise<IWeekDailySteps> =>
   new Promise(resolve => {
     RNFitnessTracker.getDailyWeekData((data: IWeekDailySteps) => {
       resolve(data);
@@ -142,7 +129,7 @@ export const getWeeklySteps = (): Promise<IWeekDailySteps> =>
  * on `iOS simulator` returns mock data
  * @return {Promise<IStepTrackerData>}
  */
-export const getSteps = (): Promise<IStepTrackerData> =>
+const getSteps = (): Promise<IStepTrackerData> =>
   new Promise(resolve => {
     // Return mock data if device is iOS simulator
     if (global.isIOS && DeviceInfo.isEmulatorSync()) {
@@ -168,3 +155,15 @@ export const getSteps = (): Promise<IStepTrackerData> =>
       });
     }
   });
+
+export const FitnessTrackerAPI = {
+  getSteps,
+  getStepsThisWeek,
+  getStepsToday,
+  getWeeklySteps,
+  isStepTrackingAvailable,
+  isStepTrackingAvailableAndroid,
+  isStepTrackingAvailableIOS,
+  isStepTrackingSupported,
+  setupStepTracking,
+};
